backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aero-kit-check-backend"
version = "0.0.1"
description = "Aero Kit Check Backend API (no-ML, no-infra skeleton in 1st version of backend)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "SQLAlchemy>=2.0.32",
    "alembic>=1.13.2",
    "asyncpg>=0.29.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "bcrypt>=4.0.0,<5.0.0",
    "PyJWT>=2.9",
    "psycopg2-binary>=2.9.0",
    # ML dependencies
    "ultralytics>=8.3.0",  # YOLO v11
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python-headless>=4.8.0",  # Headless version for Docker
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0",
    "ruff>=0.4.10",
    "mypy>=1.10.0"
]

[tool.ruff]
line-length = 100

[tool.setuptools]
package-dir = {"" = "src"}
packages = { find = { where = ["src"] } }